Split factory tests
